breaking changes: update interfaces and generic classes
(jpa) entities' identifiers must be Long
(api) never communicate database identifiers to consumers
(api) controllers may fetch individual items by different means

+++ src/main/java/org/trebol/api/services/PublicProductsService.java
@@ -17,9 +17,9 @@ public interface PublicProductsService {
       Map<String, String> allRequestParams);
 
   @Nullable
-  public ProductPojo getProduct(Integer id);
+  public ProductPojo getProduct(long id);
 
   public Collection<ProductCategoryPojo> getRootCategories();
 
-  public Collection<ProductCategoryPojo> getChildrenCategories(int parentId);
+  public Collection<ProductCategoryPojo> getChildrenCategories(long parentId);
 }
